Rename entity into venue in the UI layer
